attempt to change gridCount to be more flexible
didn't quite work.  The cotrol grid and Streamwall grid appeared to be setup correctly as far as visuals go (used countW=4 countH=3 in a custom .toml file) but clicking streams to load would only load the first 3 stream slots into streamwall window and the control grid was not populating with stream labels.

when i typed a stream label into the last grid input (bottom-right) every intput in the control grid filled with the same text but did not change the streamwall window

:(